decoder: Prevent signed/unsigned comparison in rlefont
Adding a signed and unsigned 16 bit value in C promotes the signed value
to unsigned, losing the sign bit in the process. Comparing then with an
signed value can cause potential bugs and at least throws a warning when
compiling when `sign-compare` is enabled.

This patch casts the int16_t first to int32_t to prevent the implicit
cast to uint16_t.

+++ decoder/mf_rlefont.c
@@ -66,7 +66,7 @@ static void write_pixels(struct renderstate_r *rstate, uint16_t count,
     uint8_t rowlen;
 
     /* Write row-by-row if the run spans multiple rows. */
-    while (rstate->x + count >= rstate->x_end)
+    while ((int32_t)rstate->x + count >= rstate->x_end)
     {
         rowlen = rstate->x_end - rstate->x;
         rstate->callback(rstate->x, rstate->y, rowlen, alpha, rstate->state);
